+++ src/components/common/MoveButton.tsx
@@ -1,4 +1,4 @@
-import { useContext } from 'react';
+import { useEffect, useState } from 'react';
 
 import { IconButtonProps } from '@mui/material/IconButton';
 
@@ -8,13 +8,21 @@ import {
   MoveButton as GraaspMoveButton,
 } from '@graasp/ui';
 
+import { hooks, mutations } from '@/config/queryClient';
+import { applyEllipsisOnLength, getDirectParentId } from '@/utils/item';
+
 import { useBuilderTranslation } from '../../config/i18n';
 import {
   ITEM_MENU_MOVE_BUTTON_CLASS,
   ITEM_MOVE_BUTTON_CLASS,
 } from '../../config/selectors';
 import { BUILDER } from '../../langs/constants';
-import { MoveItemModalContext } from '../context/MoveItemModalContext';
+import { NavigationElement } from '../main/itemSelectionModal/Breadcrumbs';
+import ItemSelectionModal, {
+  ItemSelectionModalProps,
+} from '../main/itemSelectionModal/ItemSelectionModal';
+
+const TITLE_MAX_NAME_LENGTH = 15;
 
 type MoveButtonProps = {
   itemIds: string[];
@@ -25,36 +33,113 @@ type MoveButtonProps = {
 };
 
 const MoveButton = ({
-  itemIds,
+  itemIds: defaultItemsIds,
   color = 'default',
   id,
   type = ActionButton.ICON_BUTTON,
   onClick,
-}: MoveButtonProps): JSX.Element | null => {
+}: MoveButtonProps): JSX.Element => {
   const { t: translateBuilder } = useBuilderTranslation();
+  const { mutate: moveItems } = mutations.useMoveItems();
 
-  const { openModal: openMoveModal } = useContext(MoveItemModalContext);
+  const [open, setOpen] = useState(false);
+  const [itemIds, setItemIds] = useState<string[]>(defaultItemsIds || []);
 
-  // TODO: return error?
-  if (!openMoveModal) {
-    return null;
-  }
+  const { data: items } = hooks.useItems(itemIds);
+
+  const openMoveModal = (newItemIds: string[]) => {
+    setOpen(true);
+    setItemIds(newItemIds);
+  };
+
+  const onClose = () => {
+    setOpen(false);
+  };
+
+  const onConfirm: ItemSelectionModalProps['onConfirm'] = (payload) => {
+    // change item's root id to null
+    const newPayload = {
+      ids: itemIds,
+      to: payload,
+    };
+    moveItems(newPayload);
+    onClose();
+  };
+  useEffect(() => {
+    // necessary to sync prop with a state because move-many-items' targets are updated dynamically with the table
+    setItemIds(defaultItemsIds);
+  }, [defaultItemsIds]);
 
   const handleMove = () => {
     openMoveModal(itemIds);
     onClick?.();
   };
 
+  const isDisabled = (item: NavigationElement, homeId: string) => {
+    if (items?.data) {
+      // cannot move inside self and below
+      const moveInSelf = Object.values(items.data).some((i) =>
+        item.path.includes(i.path),
+      );
+
+      // cannot move in same direct parent
+      // todo: not opti because we only have the ids from the table
+      const directParentIds = Object.values(items.data).map((i) =>
+        getDirectParentId(i.path),
+      );
+      const moveInDirectParent = directParentIds.includes(item.id);
+
+      // cannot move to home if was already on home
+      let moveToHome = false;
+
+      if (items?.data) {
+        moveToHome =
+          item.id === homeId &&
+          !getDirectParentId(Object.values(items.data)[0].path);
+      }
+      return moveInSelf || moveInDirectParent || moveToHome;
+    }
+    return false;
+  };
+
+  const title = items
+    ? translateBuilder(BUILDER.MOVE_ITEM_MODAL_TITLE, {
+        name: applyEllipsisOnLength(
+          Object.values(items.data)[0].name,
+          TITLE_MAX_NAME_LENGTH,
+        ),
+        // -1 because we show one name
+        count: itemIds.length - 1,
+      })
+    : translateBuilder(BUILDER.MOVE_ITEM_MODAL_TITLE);
+
+  const buttonText = (name?: string) =>
+    translateBuilder(BUILDER.MOVE_BUTTON, { name, count: name ? 1 : 0 });
+
   return (
-    <GraaspMoveButton
-      color={color}
-      type={type}
-      id={id}
-      onClick={handleMove}
-      text={translateBuilder(BUILDER.MOVE_BUTTON)}
-      menuItemClassName={ITEM_MENU_MOVE_BUTTON_CLASS}
-      iconClassName={ITEM_MOVE_BUTTON_CLASS}
-    />
+    <>
+      <GraaspMoveButton
+        color={color}
+        type={type}
+        id={id}
+        onClick={handleMove}
+        text={translateBuilder(BUILDER.MOVE_BUTTON)}
+        menuItemClassName={ITEM_MENU_MOVE_BUTTON_CLASS}
+        iconClassName={ITEM_MOVE_BUTTON_CLASS}
+      />
+
+      {items?.data && open && (
+        <ItemSelectionModal
+          title={title}
+          isDisabled={isDisabled}
+          buttonText={buttonText}
+          onClose={onClose}
+          open={open}
+          onConfirm={onConfirm}
+          items={Object.values(items.data)}
+        />
+      )}
+    </>
   );
 };