+++ src/components/item/sharing/ItemMembershipSelect.tsx
@@ -15,12 +15,31 @@ import {
 } from '../../../config/selectors';
 import { BUILDER } from '../../../langs/constants';
 
+/**
+ * Type of the `disabled` prop which allows to disable a subset of the options.
+ * Undefined keys are considered as not disabled.
+ */
+type DisabledMap = {
+  [key in PermissionLevel]?: boolean;
+};
+
+const defaultDisabledMap: DisabledMap = {
+  admin: false,
+  read: false,
+  write: false,
+};
+
 export type ItemMembershipSelectProps = {
   value: string;
   onChange?: SelectProps['onChange'];
   color?: SelectProps['color'];
   showLabel?: boolean;
   displayEmpty?: boolean;
+  /**
+   * This prop allows to disable the select when passed the value `true`
+   * or to disable only certain options when passed an object where the keys are the values of the options
+   */
+  disabled?: boolean | DisabledMap;
 };
 
 const ItemMembershipSelect = ({
@@ -29,6 +48,7 @@ const ItemMembershipSelect = ({
   color,
   showLabel = true,
   displayEmpty = false,
+  disabled = false,
 }: ItemMembershipSelectProps): JSX.Element => {
   const { t: translateBuilder } = useBuilderTranslation();
   const { t: enumT } = useEnumsTranslation();
@@ -36,7 +56,8 @@ const ItemMembershipSelect = ({
   const label = showLabel
     ? translateBuilder(BUILDER.ITEM_MEMBERSHIP_PERMISSION_LABEL)
     : undefined;
-
+  const disabledMap =
+    typeof disabled === 'boolean' ? defaultDisabledMap : disabled;
   useEffect(() => {
     if (permission !== value) {
       setPermission(value);
@@ -50,7 +71,9 @@ const ItemMembershipSelect = ({
       values={Object.values(PermissionLevel).map((v) => ({
         value: v,
         text: enumT(v),
+        disabled: disabledMap[v],
       }))}
+      disabled={typeof disabled === 'boolean' ? disabled : undefined}
       buildOptionId={buildPermissionOptionId}
       value={permission}
       defaultValue={permission}